+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:restouran_app/data/model/restaurant.dart';
+import 'package:restouran_app/data/model/restaurant_list.dart';
 import 'package:restouran_app/menu/detail_page.dart';
 import 'package:restouran_app/widget/splash_screen.dart';
-import 'package:restouran_app/ui/restaurant_list_page.dart';
+import 'package:restouran_app/ui/restaurant_home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -22,7 +22,7 @@ class MyApp extends StatelessWidget {
           visualDensity: VisualDensity.adaptivePlatformDensity),
       initialRoute: RestaurantListPage.routeName,
       routes: {
-        RestaurantListPage.routeName: (context) => const SplashScreen(),
+        RestaurantListPage.routeName: (context) => const RestaurantListPage(),
         RestaurantDetailPage.routeName: (context) => RestaurantDetailPage(
               restaurant:
                   ModalRoute.of(context)?.settings.arguments as Restaurant,
