+++ lib/provider/restaurant.dart
@@ -1,0 +1,46 @@
+import 'dart:async';
+
+import 'package:restouran_app/data/api/service_api.dart';
+import 'package:restouran_app/data/model/restaurant_list.dart';
+import 'package:flutter/material.dart';
+
+enum ResultState { Loading, NoData, HasData, Error }
+
+class RestaurantProvider extends ChangeNotifier {
+  final ApiService apiService;
+
+  RestaurantProvider({required this.apiService}) {
+    _fetchAllRestaurant();
+  }
+
+  late RestaurantResult _restaurantResult;
+  late ResultState _state;
+  String _message = '';
+
+  String get message => _message;
+
+  RestaurantResult get result => _restaurantResult;
+
+  ResultState get state => _state;
+
+  Future<dynamic> _fetchAllRestaurant() async {
+    try {
+      _state = ResultState.Loading;
+      notifyListeners();
+      final restaurant = await apiService.topHeadlines();
+      if (restaurant.restaurants.isEmpty) {
+        _state = ResultState.NoData;
+        notifyListeners();
+        return _message = 'Empty Data';
+      } else {
+        _state = ResultState.HasData;
+        notifyListeners();
+        return _restaurantResult = restaurant;
+      }
+    } catch (e) {
+      _state = ResultState.Error;
+      notifyListeners();
+      return _message = 'Error --> $e';
+    }
+  }
+}
