+++ lib/widget/splash_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:restouran_app/ui/restaurant_list_page.dart';
+import 'package:restouran_app/ui/restaurant_home_page.dart';
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({Key? key}) : super(key: key);
