+++ pubspec.yaml
@@ -34,6 +34,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   google_fonts: ^2.3.1
+  http: ^0.13.4
+  provider: ^6.0.2
 
 dev_dependencies:
   flutter_test:
@@ -57,8 +59,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-    - assets/local_restaurant.json
+  # assets:
+  #   - assets/local_restaurant.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
